+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:theme="@style/Theme.EbbingPlanner"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.EbbingPlanner">
