+++ feature/home/src/main/java/com/tgyuu/home/graph/main/ui/bottomsheet/OptionsBottomSheet.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.tgyuu.common.ui.clickable
-import com.tgyuu.designsystem.component.EbbingBottomSheetHeader
+import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetHeader
 import com.tgyuu.designsystem.foundation.EbbingTheme
 import com.tgyuu.domain.model.TodoSchedule
