+++ feature/home/src/main/java/com/tgyuu/home/graph/main/ui/bottomsheet/SortTypeBottomSheet.kt
@@ -10,8 +10,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.tgyuu.designsystem.component.EbbingBottomSheetHeader
-import com.tgyuu.designsystem.component.EbbingBottomSheetListItemDefault
+import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetHeader
+import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetListItemDefault
 import com.tgyuu.designsystem.component.EbbingSolidButton
 import com.tgyuu.domain.model.SortType
