+++ feature/setting/src/main/java/com/tgyuu/setting/graph/main/ui/bottomsheet/AlarmTimeBottomSheet.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.tgyuu.designsystem.component.EbbingBottomSheetHeader
+import com.tgyuu.designsystem.component.bottomsheet.EbbingBottomSheetHeader
 import com.tgyuu.designsystem.component.EbbingSolidButton
 import com.tgyuu.designsystem.component.picker.EbbingPicker
